refactor(llm): move tagging/titling/reminders use-cases out of infrastructure into application
