+++ src/application/mod.rs
@@ -11,6 +11,9 @@ pub mod intent;
 pub mod note_persistence;
 pub mod rag;
 pub mod reminders;
+pub mod reminders_extract;
+pub mod tagging;
+pub mod titling;
 pub mod tools;
 pub mod transcription_manager;
 pub mod web_search;
